fix(scripts,fixtures): the layering gate looks at the whole workspace, not just packages/
dependency-guard's scan root was the literal `packages/` directory, so
fixtures/, examples/, corpus/ and bench/ were never checked — and an
undeclared cross-package import had been sitting in
fixtures/integration-browser the whole time, breaking a rule that already
existed and simply never ran there.

The scan root is now derived from pnpm-workspace.yaml's own `packages:`
list, so the gap cannot reopen the next time a workspace root is added.
Nothing about the packages/ rules changes: a block still fails loudly if
it is missing from the layer map, and the layering, quarry, relative-escape
and zod-floor rules are untouched. Outside packages/ a package is a
CONSUMER, not a block — it sits above every layer, so any block in the map
is a legal dependency — but rules 2 and 3 bind it exactly as they bind a
block, which is what caught everything below.

The 34 violations the extended scan found:

- fixtures/integration-browser reached into fixtures/mcp-e2e with a
  relative path that escaped its own package. The shared piece is a real
  MCP SDK client that completes the door's OAuth dance, used by both the
  node leg and the browser leg, so it moves to
  @vendoai-fixtures/test-kit — the package whose stated purpose is "test
  helpers shared by more than one suite". Both suites now import it by
  name. test-kit compiles to dist/*.js, which is what makes this work
  from inside a Vite config; a raw .ts behind an exports map would be
  externalized and Node would choke on it.

- fixtures/redteam declared all ten of its @vendoai/* edges as
  devDependencies while importing them from src. examples/ai-sdk-agent
  (core, store), examples/demo-bank (apps, automations),
  examples/mastra-agent (store) and corpus/hosts/express-host (store) did
  the same for their test-only imports. All promoted to dependencies —
  rule 3 exists because a devDep edge is invisible, and a test-only
  import is still an import.

- corpus/harness imported @vendoai/ui in e2e-prep/overlay-mount.ts and
  never declared it. Real undeclared dependency; added.

- fixtures/test-kit declares "ai" as a peer and imports ai/test at
  runtime, so rule 4's zod floor applies: ai imports zod/v4 and a
  consumer with an older pinned zod has no ./v4 export. Added
  ">=3.25.0 <5" alongside the @modelcontextprotocol/sdk peer the new
  module needs.

- three matches in corpus/harness were import-shaped STRINGS, not
  imports: the Next layout source the corpus harness generates for the
  scraped app. Named as APP_THEME_PATH / SRC_APP_THEME_PATH and
  interpolated, which is the remedy this script's own docstring
  prescribes and reads better than the literals it replaces. The
  docstring's residual-gaps note now names generated source, not just
  comments.

No allow-list entry was added and no rule was loosened.

Generated output (.next, .turbo, coverage, test-results) joins
node_modules and dist in the skip set: a bundled chunk sitting inside a
scanned example is full of import-shaped strings and would fail the gate
on code nobody wrote. All are gitignored and nothing tracked lives under
them; .vendo/, which IS tracked and does hold source, is deliberately
absent.

No changeset: every package touched is private and unpublished, and
`changeset status --since=origin/main` confirms nothing to bump.

Proven, not claimed: violations planted in fixtures/, examples/, corpus/
and bench/ each failed `pnpm lint`, a control planted in packages/guard
still tripped the untouched layering rule, and all were reverted. The
relocated helper is exercised end to end by the full mcp-e2e suite (19
passed, 5 skipped) and the integration-browser Playwright leg (6 passed
in a real Chromium), plus corpus/harness's e2e-prep suite (11 passed) for
the string rename.

Tandem: no-op for vendo-web. This is repo-internal tooling — a lint
script's scan root and private fixture/example manifests. Nothing this
repo emits for the console to read changed.
