+++ examples/demo-bank/package.json
@@ -22,6 +22,8 @@
     "@radix-ui/react-tabs": "^1.1.15",
     "@radix-ui/react-tooltip": "^1.2.10",
     "@vendoai/actions": "workspace:*",
+    "@vendoai/apps": "workspace:*",
+    "@vendoai/automations": "workspace:*",
     "@vendoai/core": "workspace:*",
     "@vendoai/guard": "workspace:*",
     "@vendoai/harnesses": "workspace:*",
@@ -51,8 +53,6 @@
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
-    "@vendoai/apps": "workspace:*",
-    "@vendoai/automations": "workspace:*",
     "eslint": "^9",
     "eslint-config-next": "16.2.9",
     "jsdom": "^29.1.1",
