+++ examples/mastra-agent/package.json
@@ -21,6 +21,7 @@
     "@mastra/memory": "^1.22.2",
     "@mastra/observability": "^1.16.0",
     "@vendoai/core": "workspace:*",
+    "@vendoai/store": "workspace:*",
     "@vendoai/ui": "workspace:*",
     "@vendoai/vendo": "workspace:*",
     "ai": "6.0.230",
@@ -36,7 +37,6 @@
     "@types/react": "^19.2.0",
     "@types/react-dom": "^19.2.0",
     "@vendoai-fixtures/test-kit": "workspace:*",
-    "@vendoai/store": "workspace:*",
     "eslint": "^9",
     "eslint-config-next": "16.2.9",
     "tailwindcss": "^4",
