+++ fixtures/test-kit/package.json
@@ -5,6 +5,10 @@
   "private": true,
   "type": "module",
   "exports": {
+    "./mcp-client": {
+      "types": "./dist/mcp-client.d.ts",
+      "default": "./dist/mcp-client.js"
+    },
     "./node-bridge": {
       "types": "./dist/node-bridge.d.ts",
       "default": "./dist/node-bridge.js"
@@ -19,9 +23,12 @@
     "typecheck": "tsc -p tsconfig.json --noEmit"
   },
   "peerDependencies": {
-    "ai": ">=6.0.0 <7"
+    "@modelcontextprotocol/sdk": ">=1.29.0 <2",
+    "ai": ">=6.0.0 <7",
+    "zod": ">=3.25.0 <5"
   },
   "devDependencies": {
+    "@modelcontextprotocol/sdk": "^1.29.0",
     "@types/node": "^22.0.0",
     "ai": "6.0.28",
     "typescript": "^5.6.0"
